Mostly done
The table is filled in, and every data attribute has been added to it. The table now changes based on whether it is a small enough screen. The days change if they are in February. Small changes to the general look of everything, including shadows, font sizes, general typography, etc.